feat(home): refactor homepage with hero section component and update typography system
